% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_function.R
\name{fuzzy_and}
\alias{fuzzy_and}
\title{Fuzzy And Function}
\usage{
fuzzy_and(inputA, inputB)
}
\arguments{
\item{inputA}{First input vector}

\item{inputB}{Second input vector. Must be of the same dimension as inputA.}
}
\value{
Returns the Fuzzy AND of two input values in a vector.
}
\description{
Fuzzy And Function
}
\examples{
fuzzy_and(0, -1) # = -1
fuzzy_and(0, 1) # = 0
fuzzy_and(1, 2) # = 1
fuzzy_and(1, 1) # = 1
fuzzy_and(c(0.5, 0.75), c(1.5, 1)) # = c(0.5,0.75)
}
