% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_function.R
\name{fuzzy_norm}
\alias{fuzzy_norm}
\title{Fuzzy Norm}
\usage{
fuzzy_norm(input)
}
\arguments{
\item{input}{The input (vector) data observation}
}
\value{
Returns the Fuzzy norm results of input values
}
\description{
Fuzzy Norm
}
\examples{
a = c(-1,-3,4,5)
fuzzy_norm(a) # = 13
}
