% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_function.R
\name{linalg_norm}
\alias{linalg_norm}
\title{Linear Algebra for Euclidean distance}
\usage{
linalg_norm(inputA, inputB)
}
\arguments{
\item{inputA}{First input vector}

\item{inputB}{Second input vector. Must be of the same dimension as inputA.}
}
\value{
Returns the calculation results by squares of distances between two input values
}
\description{
Linear Algebra for Euclidean distance
}
\examples{
a <- c(-3,-2,-1,3,3,2,3)
b <- c(-3,-2,-1,0,1,2,3)
linalg_norm(a,b) # = 3.605
}
