% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_function.R
\name{match_function}
\alias{match_function}
\title{Match function}
\usage{
match_function(input, category_w)
}
\arguments{
\item{input}{The input (vector) data observation}

\item{category_w}{The current category weight}
}
\value{
Returns the vector of match Sj that will be used to check the vigilance parameter
}
\description{
Match function
}
