% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedataIPM.R
\docType{data}
\name{simulatedataIPM}
\alias{simulatedataIPM}
\title{Sample Data for simulate analysis data (Using IPM 2019)}
\format{
A \code{tibble} with 27 observation as region and 4 column as variables, which are:
\describe{
\item{"AHH"}{a value of Life expectancy (Angka Harapan Hidup)}
\item{"HLS"}{a value of Expected Years of Schooling (Harapan Lama Sekolah)}
\item{"RLS"}{a value of Mean Years of Schooling (Rata-rata Lama Sekolah)}
\item{"Pengeluaran"}{a value of Expenditure (Pengeluaran)}
}
}
\source{
\url{https://www.bps.go.id/}
}
\usage{
simulatedataIPM
}
\description{
A real data of Human Development Index (Indeks Pembangunan Manusia) of West Java, Indonesia 2019
}
\keyword{datasets}
