% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_function.R
\name{update_weight}
\alias{update_weight}
\title{Update weight}
\usage{
update_weight(input, category_w, beta)
}
\arguments{
\item{input}{The input (vector) data observation}

\item{category_w}{The current category weight}

\item{beta}{Learning rate in (0,1)}
}
\value{
Returns the updated weight
}
\description{
Update weight
}
