% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_function.R
\name{vigilance_check}
\alias{vigilance_check}
\title{Vigilance check}
\usage{
vigilance_check(input, category_w, rho)
}
\arguments{
\item{input}{The input (vector) data observation}

\item{category_w}{The current category weight}

\item{rho}{Vigilance parameter (0,1)}
}
\value{
Returns Boolean value (True or False) as a result of checking the match Sj vector passed the vigilance parameter or not
}
\description{
Vigilance check
}
